move exception handler to main server class and handle ServiceException instead of Ixmp4Error
